# backend/StigviddAPI/Controllers/TrailsController.cs
// SPDX-FileCopyrightText: 2025-2026 The Stigvidd Authors
// SPDX-License-Identifier: AGPL-3.0-or-later

using Core.Interfaces.Services;
using Core.Services;
using Microsoft.AspNetCore.Authorization;
using Microsoft.AspNetCore.Mvc;
using WebDataContracts.RequestModels.Media;
using WebDataContracts.RequestModels.Trail;
using WebDataContracts.ResponseModels.Trail;

namespace StigviddAPI.Controllers;

[ApiController]
[Route("api/v1/[controller]")]
public class TrailsController : StigViddController
{
    private readonly ITrailService _trailService;
    private readonly IUserService _userService;
    private readonly ILogger<TrailsController> _logger;

    public TrailsController(ITrailService trailService, IUserService userService, ILogger<TrailsController> logger)
    {
        _trailService = trailService;
        _userService = userService;
        _logger = logger;
    }

    [AllowAnonymous]
    [HttpGet("{identifier}")]
    public async Task<ActionResult<TrailResponse?>> GetTrailByIdentifier(
        string identifier,
        CancellationToken ctoken)
    {
        var result = await _trailService.GetTrailByIdentifierWithoutCoordinatesAsync(identifier, ctoken);

        if (!result.Success && result.Message != null)
        {
            _logger.LogInformation(
                "GetTrailByIdentifier: Trail with identifier: {identifier} not found.", identifier);

            return ToActionResult(result.Message);
        }

        return Ok(result.Value);
    }

    [AllowAnonymous]
    [HttpGet("popular")]
    public async Task<ActionResult<IReadOnlyCollection<TrailOverviewResponse?>>> GetPopularTrails(
        [FromQuery] double? latitude,
        [FromQuery] double? longitude,
        CancellationToken ctoken)
    {
        var result = await _trailService.GetPopularTrailOverviewsAsync(latitude, longitude, ctoken);

        if (!result.Success && result.Message != null)
        {
            _logger.LogInformation(
               "GetPopularTrails Failed to fetch popular trails.");

            return ToActionResult(result.Message);
        }

        return Ok(result.Value);
    }

    [AllowAnonymous]
    [HttpGet("")]
    public async Task<ActionResult<IReadOnlyCollection<TrailShortInfoResponse>>> GetAllTrails(CancellationToken ctoken)
    {
        var result = await _trailService.GetAllTrailsWithBasicInfoAsync(ctoken);

        if (!result.Success && result.Message != null)
        {
            _logger.LogInformation(
              "GetAllTrails: Failed to fetch trails.");

            return ToActionResult(result.Message);
        }

        return Ok(result.Value);
    }

    [AllowAnonymous]
    [HttpGet("{identifier}/coordinates")]
    public async Task<ActionResult<CoordinatesResponse?>> GetCoordinatesByTrailIdentifier(
        string identifier,
        CancellationToken ctoken)
    {
        var result = await _trailService.GetCoordinatesByTrailIdentifierAsync(identifier, ctoken);

        if (!result.Success && result.Message != null)
        {
            _logger.LogInformation(
              "GetAllCordsByTrailIdentifier: Failed to fetch cords. {identifier}", identifier);

            return ToActionResult(result.Message);
        }

        return Ok(result.Value);
    }

    [AllowAnonymous]
    [HttpGet("{identifier}/card")]
    public async Task<ActionResult<TrailCardResponse?>> GetTrailCard(
        string identifier,
        CancellationToken ctoken)
    {
        var result = await _trailService.GetTrailCardByIdentifierAsync(identifier, ctoken);

        if (!result.Success && result.Message != null)
        {
            _logger.LogInformation("GetTrailCard: Trail card with identifier: {identifier} not found.", identifier);
            return ToActionResult(result.Message);
        }

        return Ok(result.Value);
    }

    [AllowAnonymous]
    [HttpPost("cards")]
    public async Task<ActionResult<IReadOnlyCollection<TrailCardResponse>>> GetTrailCards(
        [FromBody] GetTrailCardsRequest request,
        CancellationToken ctoken)
    {
        var result = await _trailService.GetTrailCardsByIdentifiersAsync(request.Identifiers, ctoken);

        if (!result.Success && result.Message != null)
        {
            _logger.LogInformation("GetTrailCards: Failed to fetch trail cards.");
            return ToActionResult(result.Message);
        }

        return Ok(result.Value);
    }

    [AllowAnonymous]
    [HttpGet("markers")]
    public async Task<ActionResult<IReadOnlyCollection<TrailMarkerResponse>>> GetTrailMarkers(
               CancellationToken ctoken)
    {
        var result = await _trailService.GetAllTrailMarkersAsync(ctoken);

        if (!result.Success && result.Message != null)
        {
            _logger.LogInformation(
              "GetTrailMarkers: Failed to fetch trail markers.");

            return ToActionResult(result.Message);
        }

        return Ok(result.Value);
    }

    [Authorize(Policy = "Admin")]
    [HttpPut("{identifier}")]
    public async Task<ActionResult<TrailResponse?>> UpdateTrail(
        string identifier,
        [FromBody] UpdateTrailRequest request,
        CancellationToken ctoken)
    {
        var userResponse = await GetAuthenticatedUserAsync(_userService, ctoken);

        if (userResponse == null)
            return Unauthorized("User not found");

        var result = await _trailService.UpdateTrailAsync(request, identifier, userResponse.Identifier, ctoken);

        if (!result.Success && result.Message != null)
        {
            _logger.LogInformation(
                "UpdateTrail: Failed to update trail with identifier: {identifier}.", identifier);

            return ToActionResult(result.Message);
        }

        return Ok(result.Value);
    }

    [Authorize(Policy = "Admin")]
    [HttpDelete("images/{imageIdentifier}")]
    public async Task<ActionResult> DeleteTrailImage(
        string imageIdentifier,
        CancellationToken ctoken)
    {
        var userResponse = await GetAuthenticatedUserAsync(_userService, ctoken);

        if (userResponse == null)
            return Unauthorized("User not found");

        var result = await _trailService.DeleteTrailImageAsync(imageIdentifier, ctoken);

        if (!result.Success && result.Message != null)
        {
            _logger.LogInformation(
                "DeleteTrailImage: Failed to delete image with identifier: {imageIdentifier}.", imageIdentifier);

            return ToActionResult(result.Message);
        }

        return NoContent();
    }

    [Authorize(Policy = "Admin")]
    [HttpPost("{identifier}/images")]
    public async Task<ActionResult<IReadOnlyCollection<TrailImageResponse>>> AddTrailImages(
        string identifier,
        [FromForm] IFormFileCollection images,
        [FromForm] ImageProcessingOptionsRequest options,
        CancellationToken ctoken)
    {
        var userResponse = await GetAuthenticatedUserAsync(_userService, ctoken);

        if (userResponse == null)
            return Unauthorized("User not found");

        var result = await _trailService.AddTrailImagesAsync(identifier, images, options.ToOptions(), ctoken);

        if (!result.Success && result.Message != null)
        {
            _logger.LogInformation(
                "AddTrailImages: Failed to add images to trail with identifier: {identifier}.", identifier);

            return ToActionResult(result.Message);
        }

        return Ok(result.Value);
    }

    [Authorize(Policy = "Admin")]
    [HttpPost("{identifier}/symbol")]
    public async Task<ActionResult<string>> SetTrailSymbol(
        string identifier,
        [FromForm] IFormFile symbol,
        [FromForm] ImageProcessingOptionsRequest options,
        CancellationToken ctoken)
    {
        var userResponse = await GetAuthenticatedUserAsync(_userService, ctoken);

        if (userResponse == null)
            return Unauthorized("User not found");

        var result = await _trailService.SetTrailSymbolAsync(identifier, symbol, options.ToOptions(), ctoken);

        if (!result.Success && result.Message != null)
        {
            _logger.LogInformation(
                "SetTrailSymbol: Failed to set symbol for trail with identifier: {identifier}.", identifier);

            return ToActionResult(result.Message);
        }

        return Ok(new { symbolUrl = result.Value });
    }

    [Authorize(Policy = "Admin")]
    [HttpPost]
    [Route("create")]
    public async Task<ActionResult> AddTrail(
        [FromForm] CreateTrailRequest request,
        [FromForm] IFormFile trailSymbolImage,
        [FromForm] IFormFileCollection images,
        CancellationToken ctoken)
    {
        var userResponse = await GetAuthenticatedUserAsync(_userService, ctoken);

        if (userResponse == null)
        {
            return Unauthorized("User not found");
        }

        var result = await _trailService.AddTrailAsync(request, trailSymbolImage, images, userResponse.Identifier, ctoken);

        if (!result.Success && result.Message != null)
        {
            return ToActionResult(result.Message);
        }

        if (result.Value == null)
        {
            return BadRequest("Trail could not be created.");
        }

        return Created($"/api/v1/trail/{result.Value.Identifier}", result.Value);
    }
}