# web/openapi.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "StigVidd",
    "description": "Stigvidd API. This is free software, licensed under the GNU Affero General Public License version 3 or later. Under section 13 of that licence you are entitled to the Corresponding Source of this running service: https://github.com/themalind/stigvidd",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost"
    }
  ],
  "paths": {
...
    "/api/v1/Reviews/trail/{trailIdentifier}": {
      "get": {
        "tags": [
          "Reviews"
        ],
        "operationId": "Reviews_GetReviewsByTrailIdentifier",
        "parameters": [
          {
            "name": "trailIdentifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 2
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 3
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedReviewResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
    "/api/v1/TrailObstacles/trail/{trailIdentifier}": {
      "get": {
        "tags": [
          "TrailObstacles"
        ],
        "operationId": "TrailObstacles_GetTrailObstacles",
        "parameters": [
          {
            "name": "trailIdentifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TrailObstacleResponse"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
    "/api/v1/Trails/{identifier}": {
      "get": {
        "tags": [
          "Trails"
        ],
        "operationId": "Trails_GetTrailByIdentifier",
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrailResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "Trails"
        ],
        "operationId": "Trails_UpdateTrail",
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTrailRequest"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrailResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/Trails/popular": {
      "get": {
        "tags": [
          "Trails"
        ],
        "operationId": "Trails_GetPopularTrails",
        "parameters": [
          {
            "name": "latitude",
            "in": "query",
            "schema": {
              "type": "number",
              "format": "double",
              "nullable": true
            },
            "x-position": 1
          },
          {
            "name": "longitude",
            "in": "query",
            "schema": {
              "type": "number",
              "format": "double",
              "nullable": true
            },
            "x-position": 2
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TrailOverviewResponse"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/Trails": {
      "get": {
        "tags": [
          "Trails"
        ],
        "operationId": "Trails_GetAllTrails",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TrailShortInfoResponse"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/Trails/{identifier}/coordinates": {
      "get": {
        "tags": [
          "Trails"
        ],
        "operationId": "Trails_GetCoordinatesByTrailIdentifier",
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoordinatesResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/Trails/{identifier}/card": {
      "get": {
        "tags": [
          "Trails"
        ],
        "operationId": "Trails_GetTrailCard",
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrailCardResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/Trails/cards": {
      "post": {
        "tags": [
          "Trails"
        ],
        "operationId": "Trails_GetTrailCards",
        "requestBody": {
          "x-name": "request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetTrailCardsRequest"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TrailCardResponse"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/Trails/markers": {
      "get": {
        "tags": [
          "Trails"
        ],
        "operationId": "Trails_GetTrailMarkers",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TrailMarkerResponse"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/Trails/images/{imageIdentifier}": {
      "delete": {
        "tags": [
          "Trails"
        ],
        "operationId": "Trails_DeleteTrailImage",
        "parameters": [
          {
            "name": "imageIdentifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/Trails/{identifier}/images": {
      "post": {
        "tags": [
          "Trails"
        ],
        "operationId": "Trails_AddTrailImages",
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "properties": {
                  "images": {
                    "type": "array",
                    "nullable": true,
                    "items": {}
                  },
                  "MaxWidth": {
                    "type": "integer",
                    "format": "int32",
                    "nullable": true
                  },
                  "MaxHeight": {
                    "type": "integer",
                    "format": "int32",
                    "nullable": true
                  },
                  "Quality": {
                    "type": "integer",
                    "format": "int32",
                    "nullable": true
                  },
                  "Format": {
                    "type": "string",
                    "nullable": true
                  },
                  "CropX": {
                    "type": "integer",
                    "format": "int32",
                    "nullable": true
                  },
                  "CropY": {
                    "type": "integer",
                    "format": "int32",
                    "nullable": true
                  },
                  "CropWidth": {
                    "type": "integer",
                    "format": "int32",
                    "nullable": true
                  },
                  "CropHeight": {
                    "type": "integer",
                    "format": "int32",
                    "nullable": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TrailImageResponse"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/Trails/{identifier}/symbol": {
      "post": {
        "tags": [
          "Trails"
        ],
        "operationId": "Trails_SetTrailSymbol",
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "properties": {
                  "ContentType": {
                    "type": "string",
                    "nullable": true
                  },
                  "ContentDisposition": {
                    "type": "string",
                    "nullable": true
                  },
                  "Headers": {
                    "type": "array",
                    "nullable": true,
                    "items": {}
                  },
                  "Length": {
                    "type": "integer",
                    "format": "int64"
                  },
                  "Name": {
                    "type": "string",
                    "nullable": true
                  },
                  "FileName": {
                    "type": "string",
                    "nullable": true
                  },
                  "MaxWidth": {
                    "type": "integer",
                    "format": "int32",
                    "nullable": true
                  },
                  "MaxHeight": {
                    "type": "integer",
                    "format": "int32",
                    "nullable": true
                  },
                  "Quality": {
                    "type": "integer",
                    "format": "int32",
                    "nullable": true
                  },
                  "Format": {
                    "type": "string",
                    "nullable": true
                  },
                  "CropX": {
                    "type": "integer",
                    "format": "int32",
                    "nullable": true
                  },
                  "CropY": {
                    "type": "integer",
                    "format": "int32",
                    "nullable": true
                  },
                  "CropWidth": {
                    "type": "integer",
                    "format": "int32",
                    "nullable": true
                  },
                  "CropHeight": {
                    "type": "integer",
                    "format": "int32",
                    "nullable": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/Trails/create": {
      "post": {
        "tags": [
          "Trails"
        ],
        "operationId": "Trails_AddTrail",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "properties": {
                  "Name": {
                    "type": "string",
                    "nullable": true
                  },
                  "TrailLength": {
                    "type": "number",
                    "format": "decimal"
                  },
                  "Classification": {
                    "type": "integer",
                    "format": "int32",
                    "nullable": true
                  },
                  "Accessibility": {
                    "type": "boolean",
                    "nullable": true
                  },
                  "AccessibilityInfo": {
                    "type": "string",
                    "nullable": true
                  },
                  "TrailSymbol": {
                    "type": "string",
                    "nullable": true
                  },
                  "TrailSymbolImage": {
                    "type": "string",
                    "nullable": true
                  },
                  "Description": {
                    "type": "string",
                    "nullable": true
                  },
                  "FullDescription": {
                    "type": "string",
                    "nullable": true
                  },
                  "Coordinates": {
                    "type": "string"
                  },
                  "Tags": {
                    "type": "string",
                    "nullable": true
                  },
                  "IsVerified": {
                    "type": "boolean",
                    "nullable": true
                  },
                  "City": {
                    "type": "string",
                    "nullable": true
                  },
                  "ContentType": {
                    "type": "string",
                    "nullable": true
                  },
                  "ContentDisposition": {
                    "type": "string",
                    "nullable": true
                  },
                  "Headers": {
                    "type": "array",
                    "nullable": true,
                    "items": {}
                  },
                  "Length": {
                    "type": "integer",
                    "format": "int64"
                  },
                  "FileName": {
                    "type": "string",
                    "nullable": true
                  },
                  "images": {
                    "type": "array",
                    "nullable": true,
                    "items": {}
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PagedReviewResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "reviews"
        ],
        "properties": {
          "reviews": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReviewResponse"
            }
          },
          "hasMore": {
            "type": "boolean"
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "total": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        }
      },
...
      "TrailObstacleResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "identifier",
          "description",
          "issueType"
        ],
        "properties": {
          "identifier": {
            "type": "string"
          },
          "userIdentifier": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string"
          },
          "issueType": {
            "type": "string"
          },
          "incidentLongitude": {
            "type": "number",
            "format": "decimal",
            "nullable": true
          },
          "incidentLatitude": {
            "type": "number",
            "format": "decimal",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "solvedVotes": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/TrailObstacleSolvedVoteResponse"
            }
          }
        }
      },
...
      "TrailResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "identifier"
        ],
        "properties": {
          "identifier": {
            "type": "string"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "trailLength": {
            "type": "number",
            "format": "decimal",
            "nullable": true
          },
          "classification": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "accessibility": {
            "type": "boolean",
            "nullable": true
          },
          "accessibilityInfo": {
            "type": "string",
            "nullable": true
          },
          "trailSymbol": {
            "type": "string",
            "nullable": true
          },
          "trailSymbolImage": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "fullDescription": {
            "type": "string",
            "nullable": true
          },
          "coordinates": {
            "type": "string",
            "nullable": true
          },
          "tags": {
            "type": "string",
            "nullable": true
          },
          "createdBy": {
            "type": "string",
            "nullable": true
          },
          "isVerified": {
            "type": "boolean"
          },
          "city": {
            "type": "string",
            "nullable": true
          },
          "visitorInformation": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/VisitorInformationResponse"
              }
            ]
          },
          "trailImagesResponse": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/TrailImageResponse"
            }
          },
          "trailLinksResponse": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/TrailLinkResponse"
            }
          }
        }
      },
...
      "TrailOverviewResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "identifier"
        ],
        "properties": {
          "identifier": {
            "type": "string"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "trailLength": {
            "type": "number",
            "format": "decimal"
          },
          "averageRating": {
            "type": "number",
            "format": "decimal"
          },
          "trailImagesResponse": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/TrailImageResponse"
            }
          }
        }
      },
...
      "CoordinatesResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "coordinates"
        ],
        "properties": {
          "coordinates": {
            "type": "string"
          }
        }
      },
...
      "GetTrailCardsRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "identifiers"
        ],
        "properties": {
          "identifiers": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "UpdateTrailRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "name",
          "trailLength"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "trailLength": {
            "type": "number",
            "format": "decimal"
          },
          "classification": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "accessibility": {
            "type": "boolean",
            "nullable": true
          },
          "accessibilityInfo": {
            "type": "string",
            "nullable": true
          },
          "trailSymbol": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "fullDescription": {
            "type": "string",
            "nullable": true
          },
          "tags": {
            "type": "string",
            "nullable": true
          },
          "city": {
            "type": "string",
            "nullable": true
          },
          "visitorInformation": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/UpdateVisitorInformationRequest"
              }
            ]
          }
        }
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Skriv: Bearer {din Keycloak access token}",
        "name": "Authorization",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}